Use exact album matches instead of searching
The album search is weirdly inconsistent when it comes to matching
titles, often matching multiple albums and the results are not sorted by
match quality, so there's a chance that the rating could go to the wrong
track.

Using exact string matches appears to be both more accurate and quicker
(since one request for all albums is performed, instead of many
individual album searches).